Created Menu class, 3 methods and their tests.
Created method 'initialize'  wich contain the instant variable @list which
is equal to an array of arrays wich contains two item, a food and its price.
Created method 'show_menu' which puts the list of all the food in @list and
their prices.
Created 'select_dishes' method. It accept infinite argument wich should
to be fix num. The numbers will be used to choose the food from the
@list array.  Then the method puts the names and prices of the selected
foods.
Created 'payment_message' wich in base of the dishes selected from the
previous method shows up the sum of the prices. Then return a message
telling how much the user should be pay.

-All tests green
-Coverage 100%
-Rubocop Happy
-3 Domain model tables created.