Reorganized the order system with an interface user friendly.
Created a new method 'food_and_quantity' which print in the terminal
an interface that ask to the user which dish want to order and how many
portions. Then the method 'select_dishes' has been reorganized. It prints
out what the user selcted. The method payment message prints out
how much the user have to pay.
New method 'user_payment' created. It ask to the user to insert
the amount of money to pay the order. If it is different than
what should be paid  an error message is raised by the method
'payment_error'.

Created tests for each methods. They check if the methods print the correct
information.

-Test all green
-Coverage 100%

Explain the problem that this commit is solving. Focus on why you
are making this change as opposed to how (the code explains that).
Are there side effects or other unintuitive consequences of this
change? Here's the place to explain them.

Further paragraphs come after blank lines.

 - Bullet points are okay, too

 - Typically a hyphen or asterisk is used for the bullet, preceded
   by a single space, with blank lines in between, but conventions
   vary here

If you use an issue tracker, put references to them at the bottom,
like this:

Resolves: #123
See also: #456, #789