Make final changes to improve code
ListOfDishes - The initiazation argument was simplified, so it now
expects an array rather than a hash. An error check was also added
that prevents dishes with duplicate names.

Order - The order phone number was added as an instance variable
and passed to the send_text method

TextHandler - send_confirmation_text now accepts an argument
specifying the phone number to call.