Improve TextHandler and its tests further
The call to the Twilio text client has been wrapped in a
begin/rescue/end block. This is because a call to the client raises
an error if it fails, and the rescue catches that error and returns
a more meaningful error specific to this application.

It also allowed a specific test to be written that tests for the
method raising an error and that the appropriate error message is
generated.